Move LocationTest to single file, fail tests
